Fixing signedness in glib/goptions.c
glib/goption.c: In function ‘context_has_h_entry’:
glib/goption.c:785:21: error: comparison of integer expressions of different signedness: ‘gsize’ {aka ‘long unsigned int’} and ‘gint’ {aka ‘int’} [-Werror=sign-compare]
       for (i = 0; i < context->main_group->n_entries; i++)
                     ^
glib/goption.c:797:21: error: comparison of integer expressions of different signedness: ‘gsize’ {aka ‘long unsigned int’} and ‘gint’ {aka ‘int’} [-Werror=sign-compare]
       for (i = 0; i < group->n_entries; i++)
                     ^

+++ glib/goption.c
@@ -777,7 +777,7 @@ group_list_has_visible_entries (GOptionContext *context,
 static gboolean
 context_has_h_entry (GOptionContext *context)
 {
-  gsize i;
+  gint i;
   GList *list;
 
   if (context->main_group)
